fix step parsing

+++ unituring.py
@@ -172,7 +172,7 @@ def main():
             elif argument == '-d':
                 delimiter = arguments.pop(0)
             elif argument == '-s':
-                step = int(arguments.pop(0))
+                step = float(arguments.pop(0))
             elif argument == '-t':
                 tape = load_tape(arguments.pop(0), delim=delimiter)
             elif argument == '-v':
